@@ -25,79 +25,97 @@
   "config": {
       "platform": {
         "php": "8.1",
-        "ext-mongodb": "1.15.0",
         "ext-gd": "1",
         "ext-dom": "1",
-        "ext-pdo": "1",
+        "ext-gnupg": "1",
         "ext-json": "1",
+        "ext-mongodb": "2.0.0",
+        "ext-pdo": "1",
         "ext-soap": "1",
-        "ext-zip": "1",
-        "ext-fileinfo": "1"
+        "ext-sockets": "1",
+        "ext-zip": "1"
       }
   },
   "minimum-stability": "dev",
   "prefer-stable": true,
   "require": {
     "php": ">=8.1",
+    "ext-ctype": "*",
     "ext-dom": "*",
-    "ext-pdo": "*",
+    "ext-filter": "*",
+    "ext-gd": "*",
+    "ext-gnupg": "*",
     "ext-json": "*",
-    "ext-soap": "*",
-    "ext-zip": "*",
+    "ext-libxml": "*",
+    "ext-mongodb": "*",
     "ext-openssl": "*",
-    "ext-fileinfo": "*",
-    "slim/slim": "^4.3",
-    "slim/http": "^1.2",
-    "slim/flash": "^0.4",
-    "slim/twig-view": "3.3.0",
-    "nyholm/psr7": "^1.3",
-    "nyholm/psr7-server": "^1.0.1",
-    "erusev/parsedown": "~1.5",
-    "league/oauth2-server": "^8.0",
-    "flynsarmy/slim-monolog": "~1.0",
-    "sallar/jdatetime": "dev-master",
-    "gettext/gettext": "~4.0",
-    "twig/twig": "3.11.*",
-    "respect/validation": "2.2.*",
-    "intervention/image": "2.7.2",
-    "phpmailer/phpmailer": "6.5.*",
-    "guzzlehttp/guzzle": "7.5.*",
-    "league/oauth2-client": "^2.4",
-    "james-heinrich/getid3": "^1.9",
-    "onelogin/php-saml": "4.1.*",
-    "infostars/picofeed": "dev-westphal/php8",
-    "tedivm/stash": "^v0.17.6",
-    "phenx/php-font-lib": "^0.5.0",
-    "symfony/event-dispatcher": "^4.1",
-    "xibosignage/support": "0.*",
-    "xibosignage/oauth2-xibo-cms": "dev-feature/3.0",
-    "mtdowling/cron-expression": "^1.1",
-    "ralouphie/mimey": "^1.0",
-    "robmorgan/phinx": "0.13.*",
-    "johngrogg/ics-parser": "^3.1",
-    "phpseclib/mcrypt_compat": "^1.0",
-	"psr/log": "1.1.*",
-    "mongodb/mongodb": "^1.10",
-    "apereo/phpcas": "1.6.*",
-	"robthree/twofactorauth": "1.8.2",
-    "mpdf/mpdf": "v8.1.*",
-    "php-di/php-di" : "7.0.*",
-    "php-di/slim-bridge" : "3.3.*",
+    "ext-pdo": "*",
+    "ext-session": "*",
+    "ext-soap": "*",
+    "ext-sockets": "*",
+    "ext-simplexml": "*",
+    "ext-zip": "*",
     "akrabat/ip-address-middleware" : "2.1.0",
-    "selective/samesite-cookie": "0.3.0",
-    "nesbot/carbon": "2.66.*",
+    "apereo/phpcas": "1.6.*",
+    "dragonmantank/cron-expression": "^3.4",
+    "erusev/parsedown": "~1.5",
+    "flynsarmy/slim-monolog": "~1.0",
+    "gettext/gettext": "~4.0",
+    "guzzlehttp/guzzle": "7.9.*",
+    "guzzlehttp/psr7": "2.7.*",
+    "illuminate/support": "v10.48.*",
+    "infostars/picofeed": "dev-westphal/php8",
+    "intervention/image": "2.7.2",
+    "james-heinrich/getid3": "1.9.*",
     "jmikola/geojson": "^1.0",
+    "johngrogg/ics-parser": "^3.1",
+    "league/oauth2-server": "^8.0",
+    "lcobucci/clock": "3.0.*",
+    "lcobucci/jwt": "5.3.*",
+    "mjaschen/phpgeo": "^5.0",
+    "mongodb/mongodb": "2.0.*",
+    "monolog/monolog": "2.10.*",
+    "mpdf/mpdf": "v8.1.*",
+    "nesbot/carbon": "^2.73.0",
+    "nyholm/psr7": "^1.8",
+    "onelogin/php-saml": "4.1.*",
+    "phenx/php-font-lib": "^0.5.0",
+    "php-di/php-di": "7.0.*",
+    "php-di/slim-bridge": "3.3.*",
+    "phpmailer/phpmailer": "6.5.*",
+    "psr/container": "2.0.*",
+    "psr/http-message": "1.1.*",
+    "psr/http-server-handler": "1.0.*",
+    "psr/http-server-middleware": "1.0.*",
+    "psr/log": "1.1.*",
+    "ralouphie/mimey": "^1.0",
+    "respect/validation": "2.2.*",
+    "robmorgan/phinx": "0.13.*",
+    "robthree/twofactorauth": "1.8.2",
+    "slim/flash": "^0.4",
+    "slim/http": "^1.2",
+    "slim/slim": "^4.14",
+    "slim/twig-view": "3.3.0",
+    "symfony/event-dispatcher": "^4.1",
+    "symfony/event-dispatcher-contracts": "^1.10",
+    "symfony/filesystem": "v6.4.*",
+    "symfony/html-sanitizer": "v6.4.*",
     "symfony/yaml": "5.4.*",
-    "mjaschen/phpgeo": "^5.0"
+    "tedivm/stash": "v1.2.*",
+    "twig/twig": "3.11.*",
+    "xibosignage/support": "dev-php8"
   },
   "require-dev": {
-    "phpunit/phpunit": "10.0.*",
-    "zircote/swagger-php": "^2",
+    "exussum12/coverage-checker": "^0.11.2",
     "doctrine/instantiator": "1.5.0",
     "doctrine/annotations": "^1",
+    "league/oauth2-client": "^2.4",
+    "micheh/phpcs-gitlab": "^1.1",
+    "phpunit/phpunit": "10.0.*",
+    "shipmonk/composer-dependency-analyser": "^1.8",
     "squizlabs/php_codesniffer": "3.*",
-    "exussum12/coverage-checker": "^0.11.2",
-    "micheh/phpcs-gitlab": "^1.1"
+    "xibosignage/oauth2-xibo-cms": "dev-feature/3.0",
+    "zircote/swagger-php": "^2"
   },
   "autoload": {
     "psr-4": { "Xibo\\" : "lib/", "Xibo\\Custom\\": "custom/" }
