@@ -1,8 +1,8 @@
 <?php
-/**
- * Copyright (C) 2020 Xibo Signage Ltd
+/*
+ * Copyright (C) 2025 Xibo Signage Ltd
  *
- * Xibo - Digital Signage - http://www.xibo.org.uk
+ * Xibo - Digital Signage - https://xibosignage.com
  *
  * This file is part of Xibo.
  *
@@ -61,7 +61,7 @@ class CampaignLayoutUnassignTest extends LocalWebTestCase
     {
         parent::setup();
 
-        $this->getLogger()->debug('Setup test for Cache ' . get_class() .' Test');
+        $this->getLogger()->debug('Setup test for Cache ' . get_class($this) .' Test');
 
         // Create a Layout
         $this->layout = $this->createLayout();
