@@ -1,8 +1,8 @@
 <?php
-/**
- * Copyright (C) 2020 Xibo Signage Ltd
+/*
+ * Copyright (C) 2025 Xibo Signage Ltd
  *
- * Xibo - Digital Signage - http://www.xibo.org.uk
+ * Xibo - Digital Signage - https://xibosignage.com
  *
  * This file is part of Xibo.
  *
@@ -67,7 +67,7 @@ class LayoutInCampaignStatusTest extends LocalWebTestCase
     {
         parent::setup();
 
-        $this->getLogger()->debug('Setup test for Cache ' . get_class() .' Test');
+        $this->getLogger()->debug('Setup test for Cache ' . get_class($this) .' Test');
 
         // Create a Campaign
         $this->campaign = (new XiboCampaign($this->getEntityProvider()))->create(Random::generateString());
