@@ -1,8 +1,8 @@
 <?php
 /*
- * Copyright (C) 2022 Xibo Signage Ltd
+ * Copyright (C) 2025 Xibo Signage Ltd
  *
- * Xibo - Digital Signage - http://www.xibo.org.uk
+ * Xibo - Digital Signage - https://xibosignage.com
  *
  * This file is part of Xibo.
  *
@@ -43,7 +43,7 @@ class InteractiveFeaturesTest extends \Xibo\Tests\LocalWebTestCase
     {
         parent::setup();
 
-        $this->getLogger()->debug('Setup test for ' . get_class());
+        $this->getLogger()->debug('Setup test for ' . get_class($this));
 
         // Create a Layout
         $this->layout = $this->createLayout();
