@@ -1,8 +1,8 @@
 <?php
-/**
- * Copyright (C) 2020 Xibo Signage Ltd
+/*
+ * Copyright (C) 2025 Xibo Signage Ltd
  *
- * Xibo - Digital Signage - http://www.xibo.org.uk
+ * Xibo - Digital Signage - https://xibosignage.com
  *
  * This file is part of Xibo.
  *
@@ -56,7 +56,7 @@ class SearchFilterTest extends LocalWebTestCase
     {
         parent::setup();
 
-        $this->getLogger()->debug('Setup test for  ' . get_class() . ' Test');
+        $this->getLogger()->debug('Setup test for  ' . get_class($this) . ' Test');
 
         // Create 6 layouts to test with
         $this->layout = (new XiboLayout($this->getEntityProvider()))
