@@ -1,8 +1,8 @@
 <?php
-/**
- * Copyright (C) 2020 Xibo Signage Ltd
+/*
+ * Copyright (C) 2025 Xibo Signage Ltd
  *
- * Xibo - Digital Signage - http://www.xibo.org.uk
+ * Xibo - Digital Signage - https://xibosignage.com
  *
  * This file is part of Xibo.
  *
@@ -56,7 +56,7 @@ class AudioWidgetTest extends LocalWebTestCase
     {
         parent::setup();
 
-        $this->getLogger()->debug('Setup for ' . get_class() .' Test');
+        $this->getLogger()->debug('Setup for ' . get_class($this) .' Test');
 
         // Create a Layout
         $this->publishedLayout = $this->createLayout();
@@ -89,7 +89,7 @@ class AudioWidgetTest extends LocalWebTestCase
 
         parent::tearDown();
 
-        $this->getLogger()->debug('Tear down for ' . get_class() .' Test');
+        $this->getLogger()->debug('Tear down for ' . get_class($this) .' Test');
     }
 
     /**
@@ -97,7 +97,7 @@ class AudioWidgetTest extends LocalWebTestCase
      */
     public function testEdit()
     {
-        $this->getLogger()->debug('testEdit ' . get_class() .' Test');
+        $this->getLogger()->debug('testEdit ' . get_class($this) .' Test');
 
         // Now try to edit our assigned Media Item.
         $name = 'Edited Name ' . Random::generateString(5);
