@@ -1,8 +1,8 @@
 <?php
-/**
- * Copyright (C) 2020 Xibo Signage Ltd
+/*
+ * Copyright (C) 2025 Xibo Signage Ltd
  *
- * Xibo - Digital Signage - http://www.xibo.org.uk
+ * Xibo - Digital Signage - https://xibosignage.com
  *
  * This file is part of Xibo.
  *
@@ -53,7 +53,7 @@ class GoogleTrafficWidgetTest extends LocalWebTestCase
     {
         parent::setup();
 
-        $this->getLogger()->debug('Setup for ' . get_class() .' Test');
+        $this->getLogger()->debug('Setup for ' . get_class($this) .' Test');
 
         // Create a Layout
         $this->publishedLayout = $this->createLayout();
@@ -77,7 +77,7 @@ class GoogleTrafficWidgetTest extends LocalWebTestCase
 
         parent::tearDown();
 
-        $this->getLogger()->debug('Tear down for ' . get_class() .' Test');
+        $this->getLogger()->debug('Tear down for ' . get_class($this) .' Test');
     }
 
 	/**
@@ -106,7 +106,7 @@ class GoogleTrafficWidgetTest extends LocalWebTestCase
      */
     public function testEdit($statusCode, $name, $duration, $useDisplayLocation, $lat, $long, $zoom)
     {
-        $this->getLogger()->debug('testEdit ' . get_class() .' Test');
+        $this->getLogger()->debug('testEdit ' . get_class($this) .' Test');
 
         $response = $this->sendRequest('PUT','/playlist/widget/' . $this->widgetId, [
             'name' => $name,
