@@ -1,8 +1,8 @@
 <?php
-/**
- * Copyright (C) 2020 Xibo Signage Ltd
+/*
+ * Copyright (C) 2025 Xibo Signage Ltd
  *
- * Xibo - Digital Signage - http://www.xibo.org.uk
+ * Xibo - Digital Signage - https://xibosignage.com
  *
  * This file is part of Xibo.
  *
@@ -53,7 +53,7 @@ class PdfWidgetTest extends LocalWebTestCase
     {
         parent::setup();
 
-        $this->getLogger()->debug('Setup for ' . get_class() .' Test');
+        $this->getLogger()->debug('Setup for ' . get_class($this) .' Test');
 
         // Create a Layout
         $this->publishedLayout = $this->createLayout();
@@ -84,7 +84,7 @@ class PdfWidgetTest extends LocalWebTestCase
 
         parent::tearDown();
 
-        $this->getLogger()->debug('Tear down for ' . get_class() .' Test');
+        $this->getLogger()->debug('Tear down for ' . get_class($this) .' Test');
     }
 
     /**
@@ -92,7 +92,7 @@ class PdfWidgetTest extends LocalWebTestCase
      */
     public function testEdit()
     {
-        $this->getLogger()->debug('testEdit ' . get_class() .' Test');
+        $this->getLogger()->debug('testEdit ' . get_class($this) .' Test');
 
         $name = 'Edited Name: ' . Random::generateString(5);
         $duration = 80;
